// sdks/typescript/core-generated/runtime-protobuf/runtime/v1/ai_scheduling.ts
export interface SchedulingOccupancySnapshot {
    /**
     * @generated from protobuf field: int32 global_used = 1
     */
    globalUsed: number;
    /**
     * @generated from protobuf field: int32 global_cap = 2
     */
    globalCap: number;
    /**
     * @generated from protobuf field: int32 app_used = 3
     */
    appUsed: number;
    /**
     * @generated from protobuf field: int32 app_cap = 4
     */
    appCap: number;
}
/**
 * K-SCHED-002: Scheduling preflight judgement result.
 *
 * @generated from protobuf message nimi.runtime.v1.SchedulingJudgement
 */
export interface SchedulingJudgement {
    /**
     * @generated from protobuf field: nimi.runtime.v1.SchedulingState state = 1
     */
    state: SchedulingState;
    /**
     * @generated from protobuf field: string detail = 2
     */
    detail: string;
    /**
     * @generated from protobuf field: nimi.runtime.v1.SchedulingOccupancySnapshot occupancy = 3
     */
    occupancy?: SchedulingOccupancySnapshot;
    /**
     * @generated from protobuf field: repeated string resource_warnings = 4
     */
    resourceWarnings: string[];
}
/**
 * K-SCHED-007: Target-scoped scheduling evaluation input.
 *
 * @generated from protobuf message nimi.runtime.v1.SchedulingResourceHint
 */
export interface SchedulingResourceHint {
    /**
     * @generated from protobuf field: int64 estimated_vram_bytes = 1
     */
    estimatedVramBytes: string;
    /**
     * @generated from protobuf field: int64 estimated_ram_bytes = 2
     */
    estimatedRamBytes: string;
    /**
     * @generated from protobuf field: int64 estimated_disk_bytes = 3
     */
    estimatedDiskBytes: string;
    /**
     * @generated from protobuf field: string engine = 4
     */
    engine: string;
}
/**
 * K-SCHED-002: Atomic scheduling evaluation target.
 *
 * @generated from protobuf message nimi.runtime.v1.SchedulingEvaluationTarget
 */
export interface SchedulingEvaluationTarget {
    /**
     * @generated from protobuf field: string capability = 1
     */
    capability: string;
    /**
     * @generated from protobuf field: string target_id = 2
     */
    targetId: string;
    /**
     * @generated from protobuf field: string profile_id = 3
     */
    profileId: string;
    /**
     * @generated from protobuf field: nimi.runtime.v1.SchedulingResourceHint resource_hint = 4
     */
    resourceHint?: SchedulingResourceHint;
}
/**
 * K-SCHED-002: Per-target scheduling judgement mapping.
 *
 * @generated from protobuf message nimi.runtime.v1.SchedulingTargetJudgement
 */
export interface SchedulingTargetJudgement {
    /**
     * @generated from protobuf field: nimi.runtime.v1.SchedulingEvaluationTarget target = 1
     */
    target?: SchedulingEvaluationTarget;
    /**
     * @generated from protobuf field: nimi.runtime.v1.SchedulingJudgement judgement = 2
     */
    judgement?: SchedulingJudgement;
}
/**
 * K-SCHED-002: PeekScheduling preflight request.
 *
 * @generated from protobuf message nimi.runtime.v1.PeekSchedulingRequest
 */
export interface PeekSchedulingRequest {
    /**
     * @generated from protobuf field: string app_id = 1
     */
    appId: string;
    /**
     * @generated from protobuf field: repeated nimi.runtime.v1.SchedulingEvaluationTarget targets = 2
     */
    targets: SchedulingEvaluationTarget[];
}
/**
 * @generated from protobuf message nimi.runtime.v1.PeekSchedulingResponse
 */
export interface PeekSchedulingResponse {
    /**
     * @generated from protobuf field: nimi.runtime.v1.SchedulingOccupancySnapshot occupancy = 1
     */
    occupancy?: SchedulingOccupancySnapshot;
    /**
     * @generated from protobuf field: nimi.runtime.v1.SchedulingJudgement aggregate_judgement = 2
     */
    aggregateJudgement?: SchedulingJudgement;
    /**
     * @generated from protobuf field: repeated nimi.runtime.v1.SchedulingTargetJudgement target_judgements = 3
     */
    targetJudgements: SchedulingTargetJudgement[];
}
/**
 * K-SCHED-001: Six-value scheduling judgement state enum.
 *
 * @generated from protobuf enum nimi.runtime.v1.SchedulingState
 */
export enum SchedulingState {
    /**
     * @generated from protobuf enum value: SCHEDULING_STATE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: SCHEDULING_STATE_RUNNABLE = 1;
     */
    RUNNABLE = 1,
    /**
     * @generated from protobuf enum value: SCHEDULING_STATE_QUEUE_REQUIRED = 2;
     */
    QUEUE_REQUIRED = 2,
    /**
     * @generated from protobuf enum value: SCHEDULING_STATE_PREEMPTION_RISK = 3;
     */
    PREEMPTION_RISK = 3,
    /**
     * @generated from protobuf enum value: SCHEDULING_STATE_SLOWDOWN_RISK = 4;
     */
    SLOWDOWN_RISK = 4,
    /**
     * @generated from protobuf enum value: SCHEDULING_STATE_DENIED = 5;
     */
    DENIED = 5,
    /**
     * @generated from protobuf enum value: SCHEDULING_STATE_UNKNOWN = 6;
     */
    UNKNOWN = 6
}
// ...
export const SchedulingOccupancySnapshot = new SchedulingOccupancySnapshot$Type();
// ...
export const SchedulingJudgement = new SchedulingJudgement$Type();
// ...
export const SchedulingResourceHint = new SchedulingResourceHint$Type();
// ...
export const SchedulingEvaluationTarget = new SchedulingEvaluationTarget$Type();
// ...
export const SchedulingTargetJudgement = new SchedulingTargetJudgement$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PeekSchedulingRequest$Type extends MessageType<PeekSchedulingRequest> {
    constructor() {
        super("nimi.runtime.v1.PeekSchedulingRequest", [
            { no: 1, name: "app_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "targets", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => SchedulingEvaluationTarget }
        ]);
    }
    create(value?: PartialMessage<PeekSchedulingRequest>): PeekSchedulingRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.appId = "";
        message.targets = [];
        if (value !== undefined)
            reflectionMergePartial<PeekSchedulingRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PeekSchedulingRequest): PeekSchedulingRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string app_id */ 1:
                    message.appId = reader.string();
                    break;
                case /* repeated nimi.runtime.v1.SchedulingEvaluationTarget targets */ 2:
                    message.targets.push(SchedulingEvaluationTarget.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PeekSchedulingRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string app_id = 1; */
        if (message.appId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.appId);
        /* repeated nimi.runtime.v1.SchedulingEvaluationTarget targets = 2; */
        for (let i = 0; i < message.targets.length; i++)
            SchedulingEvaluationTarget.internalBinaryWrite(message.targets[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.PeekSchedulingRequest
 */
export const PeekSchedulingRequest = new PeekSchedulingRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PeekSchedulingResponse$Type extends MessageType<PeekSchedulingResponse> {
    constructor() {
        super("nimi.runtime.v1.PeekSchedulingResponse", [
            { no: 1, name: "occupancy", kind: "message", T: () => SchedulingOccupancySnapshot },
            { no: 2, name: "aggregate_judgement", kind: "message", T: () => SchedulingJudgement },
            { no: 3, name: "target_judgements", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => SchedulingTargetJudgement }
        ]);
    }
    create(value?: PartialMessage<PeekSchedulingResponse>): PeekSchedulingResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.targetJudgements = [];
        if (value !== undefined)
            reflectionMergePartial<PeekSchedulingResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PeekSchedulingResponse): PeekSchedulingResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* nimi.runtime.v1.SchedulingOccupancySnapshot occupancy */ 1:
                    message.occupancy = SchedulingOccupancySnapshot.internalBinaryRead(reader, reader.uint32(), options, message.occupancy);
                    break;
                case /* nimi.runtime.v1.SchedulingJudgement aggregate_judgement */ 2:
                    message.aggregateJudgement = SchedulingJudgement.internalBinaryRead(reader, reader.uint32(), options, message.aggregateJudgement);
                    break;
                case /* repeated nimi.runtime.v1.SchedulingTargetJudgement target_judgements */ 3:
                    message.targetJudgements.push(SchedulingTargetJudgement.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PeekSchedulingResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* nimi.runtime.v1.SchedulingOccupancySnapshot occupancy = 1; */
        if (message.occupancy)
            SchedulingOccupancySnapshot.internalBinaryWrite(message.occupancy, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* nimi.runtime.v1.SchedulingJudgement aggregate_judgement = 2; */
        if (message.aggregateJudgement)
            SchedulingJudgement.internalBinaryWrite(message.aggregateJudgement, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* repeated nimi.runtime.v1.SchedulingTargetJudgement target_judgements = 3; */
        for (let i = 0; i < message.targetJudgements.length; i++)
            SchedulingTargetJudgement.internalBinaryWrite(message.targetJudgements[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.PeekSchedulingResponse
 */
export const PeekSchedulingResponse = new PeekSchedulingResponse$Type();